--- frontend/src/components/Header/Header.tsx
@@ -1,12 +1,0 @@
-import * as S from './Header.css';
-
-const Header = () => {
-  return (
-    <>
-      <header className={S.Layout}>Header</header>
-      <div className={S.Spacer} />
-    </>
-  );
-};
-
-export default Header;
