+++ frontend/src/pages/Layout/Layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Header from '@/components/Header/Header';
+import Header from '@/components/_common/Header/Header';
 import NavBar from '@/components/NavBar/NavBar';
 
 const Layout = () => {
